refactor SbTabPanel and SbTabPanels components to SFC
